Relocate JVM-only functions
