+++ runtime/runtime-core/jvm/src/aws/smithy/kotlin/runtime/util/JMESPathJVM.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package aws.smithy.kotlin.runtime.util
+
+import aws.smithy.kotlin.runtime.InternalApi
+
+@InternalApi
+@JvmName("noOpUnnestedCollection")
+public inline fun <reified T> Collection<T>.flattenIfPossible(): Collection<T> = this
+
+@InternalApi
+@JvmName("flattenNestedCollection")
+public inline fun <reified T> Collection<Collection<T>>.flattenIfPossible(): Collection<T> = flatten()
